requirements.txt
python-dotenv>=1.0.0
altair
attrs
beautifulsoup4
blinker
cachetools
certifi
cffi
charset-normalizer
click
colorama
curl_cffi
frozendict
gitdb
GitPython
idna
Jinja2
jsonschema
jsonschema-specifications
MarkupSafe
multitasking
narwhals
numpy
packaging
pandas
peewee
pillow
platformdirs
protobuf
pyarrow
pycparser
pydeck
python-dateutil
python-dotenv
pytz
referencing
requests
rpds-py
six
smmap
soupsieve
streamlit
tenacity
toml
tornado
typing_extensions
tzdata
urllib3
watchdog
websockets
yfinance
langchain
langchain-google-genai
langchain-community
duckduckgo-search
fastapi
python-multipart
